sms unit test and documentation of swagger

+++ bcb-sms-service/src/main/java/br/com/bcb/services/SmsService.java
@@ -2,6 +2,8 @@ package br.com.bcb.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import br.com.bcb.model.Sms;
@@ -9,6 +11,7 @@ import br.com.bcb.repositories.SmsRepository;
 
 @Service
 public class SmsService {
+
 	@Autowired
 	private Environment enviroment;
 	
@@ -19,6 +22,13 @@ public class SmsService {
 		return smsRepository.save(sms);
 	}
 	
+	public Page<Sms> findAll(Pageable pageable) {	
+		var smsPage = smsRepository.findAll(pageable);
+		var port = enviroment.getProperty("local.server.port");
+		smsPage.getContent().stream().forEach(c -> c.setEnviroment(port));
+		return smsPage;
+	}
+	
 	public Sms findById(Long id) {
 		var sms = smsRepository.findById(id);
 		if(sms.isEmpty()) throw new RuntimeException("Sms not found!");
@@ -27,17 +37,17 @@ public class SmsService {
 		return sms.get();
 	}
 	
-	public Sms findByClientId(Long client_id) {
-		var sms = smsRepository.findByClientId(client_id);
+	public Page<Sms> findByClientId(Long client_id, Pageable pageable) {
+		var sms = smsRepository.findByClientId(client_id, pageable);
 		if(sms.isEmpty()) throw new RuntimeException("Sms not found!");
 		var port = enviroment.getProperty("local.server.port");
-		sms.get().setEnviroment(port);
-		return sms.get();
+		sms.getContent().stream().forEach(c -> c.setEnviroment(port));
+		return sms;
 	}
 	
 	public Sms updateSms(Sms sms) {
 		var smsFound = smsRepository.findById(sms.getId());
-		if(smsFound.isEmpty()) throw new RuntimeException("Client not found!");
+		if(smsFound.isEmpty()) throw new RuntimeException("Sms not found!");
 		var smsReturn = smsFound.get();
 		var port = enviroment.getProperty("local.server.port");
 		smsReturn.setClientId(sms.getClientId());
@@ -50,6 +60,9 @@ public class SmsService {
 	}
 	
 	public void deleteSms( Long id) {
-		smsRepository.deleteById(id);;
+		var smsFound = smsRepository.findById(id);
+		if(smsFound.isEmpty()) throw new RuntimeException("Sms not found!");
+		smsRepository.deleteById(id);
 	}
+	
 }
